Add nearest neighbour interpolation routines ang2ngb.
A linear interpolation using three pixels covering the sampled
location.

The domain of gsp2ang is extended to cover infinite x and y.
in the y direction there is a reflective boundry near y = +/-pi/2,
(implying a period of 2 pi.)
in the x direction there is a periodic boundry of 2 pi.

Also extended coverage near the poles to fill the gray area
in the healpix paper Fig 5, such that the field is continuous in
near the edges of the cap. This is to make interpolation easier.